double por decimal

+++ backend/Dtos/CalculationResultDTO.cs
@@ -3,8 +3,8 @@ namespace backend.Dtos;
 public record CalculationResultDTO
 {
     public List<CoinPredictionDto> CoinPredictions { get; init; }
-    public double TotalAmount { get; set; }
-    public double TotalWeight { get; set; }
+    public decimal TotalAmount { get; set; }
+    public decimal TotalWeight { get; set; }
     public int TotalCount { get; internal set; }
 
     public CalculationResultDTO()
